requirements.txt
FastAPI~=0.61.1
SQLAlchemy~=1.3.20
pydantic~=1.6.1
gunicorn
uvicorn
uvloop
pymysql
httptools
PyYAML~=5.3.1
requests~=2.24.0
fastapi-users[sqlalchemy]
databases[mysql,sqlite]